+++ src/components/Favourites.tsx
@@ -37,7 +37,7 @@ const Favourites = ({
   return (
     <Box>
       <Typography variant="h5" sx={{ mb: 4 }}>
-        Favourites tasks
+        Favourite tasks
       </Typography>
       <Box>
         <TableContainer component={Paper} variant="outlined">
@@ -54,12 +54,11 @@ const Favourites = ({
               {favourites.map((task) => (
                 <TableRow key={task.id}>
                   <TableCell>
-                    <Tooltip
-                      title={`Project: ${task.project} | Ext.ID: ${task.external_id}`}
-                      placement="left"
-                    >
-                      <Box>{task.desc}</Box>
-                    </Tooltip>
+                    <Box>{task.desc}</Box>
+                    <Typography variant="caption" color="text.secondary">
+                      {task.project}
+                      {task.external_id && ` (${task.external_id})`}
+                    </Typography>
                   </TableCell>
                   <TableCell align="right">
                     <Tooltip title="Unmark favourite" placement="top">